+++ sal/net/ip/address.test.cpp
@@ -205,7 +205,7 @@ TEST_F(net_ip_address, store_v4)
   auto &a{reinterpret_cast<sockaddr_in &>(buf)};
 
   ASSERT_EQ(AF_INET, a.sin_family);
-  EXPECT_EQ(INADDR_LOOPBACK, ntohl(a.sin_addr.s_addr));
+  EXPECT_EQ(static_cast<uint32_t>(INADDR_LOOPBACK), ntohl(a.sin_addr.s_addr));
 }
